# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bingo OroDig API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  displayName?: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  moderator: 'moderator',
  player: 'player',
} as const;

export interface User {
  id: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserRole;
  balance: number;
  isBlocked: boolean;
  totalWins?: number;
  totalGamesPlayed?: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  moderator: 'moderator',
  player: 'player',
} as const;

export interface UserUpdate {
  displayName?: string;
  avatarUrl?: string;
  role?: UserUpdateRole;
}

export interface BalanceUpdate {
  amount: number;
  reason?: string;
}

export interface BlockToggle {
  isBlocked: boolean;
}

export type RoomType = typeof RoomType[keyof typeof RoomType];


export const RoomType = {
  classic: 'classic',
  fast: 'fast',
  vip: 'vip',
  automatic: 'automatic',
} as const;

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  active: 'active',
  inactive: 'inactive',
  playing: 'playing',
} as const;

export type RoomPatternType = typeof RoomPatternType[keyof typeof RoomPatternType];


export const RoomPatternType = {
  line: 'line',
  fullCard: 'fullCard',
  corners: 'corners',
  x: 'x',
  diagonal: 'diagonal',
} as const;

export interface Room {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  type: RoomType;
  status: RoomStatus;
  cardPrice: number;
  maxPlayers: number;
  ballInterval: number;
  prize: number;
  patternType: RoomPatternType;
  playerCount?: number;
  /** @nullable */
  currentGameId?: number | null;
  createdAt: string;
}

export type RoomInputType = typeof RoomInputType[keyof typeof RoomInputType];


export const RoomInputType = {
  classic: 'classic',
  fast: 'fast',
  vip: 'vip',
  automatic: 'automatic',
} as const;

export type RoomInputPatternType = typeof RoomInputPatternType[keyof typeof RoomInputPatternType];


export const RoomInputPatternType = {
  line: 'line',
  fullCard: 'fullCard',
  corners: 'corners',
  x: 'x',
  diagonal: 'diagonal',
} as const;

export interface RoomInput {
  name: string;
  description?: string;
  type: RoomInputType;
  cardPrice: number;
  maxPlayers: number;
  ballInterval: number;
  prize: number;
  patternType: RoomInputPatternType;
}

export type RoomUpdateType = typeof RoomUpdateType[keyof typeof RoomUpdateType];


export const RoomUpdateType = {
  classic: 'classic',
  fast: 'fast',
  vip: 'vip',
  automatic: 'automatic',
} as const;

export type RoomUpdateStatus = typeof RoomUpdateStatus[keyof typeof RoomUpdateStatus];


export const RoomUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
  playing: 'playing',
} as const;

export type RoomUpdatePatternType = typeof RoomUpdatePatternType[keyof typeof RoomUpdatePatternType];


export const RoomUpdatePatternType = {
  line: 'line',
  fullCard: 'fullCard',
  corners: 'corners',
  x: 'x',
  diagonal: 'diagonal',
} as const;

export interface RoomUpdate {
  name?: string;
  description?: string;
  type?: RoomUpdateType;
  status?: RoomUpdateStatus;
  cardPrice?: number;
  maxPlayers?: number;
  ballInterval?: number;
  prize?: number;
  patternType?: RoomUpdatePatternType;
}

export type GameStatus = typeof GameStatus[keyof typeof GameStatus];


export const GameStatus = {
  waiting: 'waiting',
  playing: 'playing',
  paused: 'paused',
  finished: 'finished',
} as const;

export type GamePatternType = typeof GamePatternType[keyof typeof GamePatternType];


export const GamePatternType = {
  line: 'line',
  fullCard: 'fullCard',
  corners: 'corners',
  x: 'x',
  diagonal: 'diagonal',
} as const;

export interface Game {
  id: number;
  roomId: number;
  status: GameStatus;
  patternType: GamePatternType;
  ballInterval: number;
  prize: number;
  /** @nullable */
  winnerId?: number | null;
  /** @nullable */
  winnerCardId?: number | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  finishedAt?: string | null;
  createdAt: string;
}

export type GameInputPatternType = typeof GameInputPatternType[keyof typeof GameInputPatternType];


export const GameInputPatternType = {
  line: 'line',
  fullCard: 'fullCard',
  corners: 'corners',
  x: 'x',
  diagonal: 'diagonal',
} as const;

export interface GameInput {
  roomId: number;
  patternType?: GameInputPatternType;
  prize?: number;
  ballInterval?: number;
}

export type GameControlAction = typeof GameControlAction[keyof typeof GameControlAction];


export const GameControlAction = {
  start: 'start',
  pause: 'pause',
  resume: 'resume',
  finish: 'finish',
  restart: 'restart',
} as const;

export interface GameControl {
  action: GameControlAction;
}

export interface DrawnNumber {
  id: number;
  gameId: number;
  number: number;
  letter: string;
  drawnAt: string;
}

export interface Winner {
  id: number;
  gameId: number;
  userId: number;
  cardId: number;
  username?: string;
  pattern: string;
  prize: number;
  createdAt: string;
}

export interface Card {
  id: number;
  userId: number;
  gameId: number;
  roomId?: number;
  /** JSON string of 5x5 grid */
  numbers: string;
  /** JSON array of marked numbers */
  markedNumbers?: string;
  isWinner: boolean;
  colorTheme?: string;
  purchasedAt: string;
}

export interface CardPurchase {
  gameId: number;
  quantity?: number;
}

export type BingoClaimPattern = typeof BingoClaimPattern[keyof typeof BingoClaimPattern];


export const BingoClaimPattern = {
  line: 'line',
  fullCard: 'fullCard',
  corners: 'corners',
  x: 'x',
  diagonal: 'diagonal',
} as const;

export interface BingoClaim {
  pattern: BingoClaimPattern;
}

export type ChatMessageType = typeof ChatMessageType[keyof typeof ChatMessageType];


export const ChatMessageType = {
  user: 'user',
  system: 'system',
  winner: 'winner',
} as const;

export interface ChatMessage {
  id: number;
  roomId: number;
  userId: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  content: string;
  type: ChatMessageType;
  createdAt: string;
}

export interface ChatInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  content: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  purchase: 'purchase',
  prize: 'prize',
} as const;

export interface Transaction {
  id: number;
  userId: number;
  type: TransactionType;
  amount: number;
  description: string;
  createdAt: string;
}

export interface DashboardStats {
  totalUsers: number;
  onlineUsers: number;
  activeGames: number;
  totalRooms: number;
  totalCardsToday: number;
  totalPrizesToday: number;
  totalRevenue: number;
  recentWinners?: Winner[];
}

export interface LeaderboardEntry {
  userId: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  totalWins: number;
  totalPrize: number;
  rank: number;
}