# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bingo OroDig API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /cards:
    get:
      operationId: listMyCards
      tags: [cards]
      summary: Get my bingo cards
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"

    post:
      operationId: buyCard
      tags: [cards]
      summary: Buy a bingo card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardPurchase"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}:
    get:
      operationId: getCard
      tags: [cards]
      summary: Get card by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}/claim:
    post:
      operationId: claimBingo
      tags: [cards]
      summary: Claim bingo on a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BingoClaim"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Winner"
        "400":
          description: Invalid claim
  /chat/{roomId}:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get chat messages for room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        displayName:
          type: string
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, role, balance, isBlocked, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, moderator, player]
        balance:
          type: number
        isBlocked:
          type: boolean
        totalWins:
          type: integer
        totalGamesPlayed:
          type: integer
        createdAt:
          type: string
...
    Winner:
      type: object
      required: [id, gameId, userId, cardId, pattern, prize, createdAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        userId:
          type: integer
        cardId:
          type: integer
        username:
          type: string
        pattern:
          type: string
        prize:
          type: number
        createdAt:
          type: string
    Card:
      type: object
      required: [id, userId, gameId, numbers, isWinner, purchasedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        gameId:
          type: integer
        roomId:
          type: integer
        numbers:
          type: string
          description: JSON string of 5x5 grid
        markedNumbers:
          type: string
          description: JSON array of marked numbers
        isWinner:
          type: boolean
        colorTheme:
          type: string
        purchasedAt:
          type: string
    CardPurchase:
      type: object
      required: [gameId]
      properties:
        gameId:
          type: integer
        quantity:
          type: integer
    BingoClaim:
      type: object
      required: [pattern]
      properties:
        pattern:
          type: string
          enum: [line, fullCard, corners, x, diagonal]
    ChatMessage:
      type: object
      required: [id, roomId, userId, username, content, type, createdAt]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        content:
          type: string
        type:
          type: string
          enum: [user, system, winner]
        createdAt:
          type: string
    ChatInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 200
...